Exchange Vip and Guard map fragments
